refactor: [#220] 회원가입 화면 파일 분리
